+++ app/assets/javascripts/components/CardApplicationSurvey/CallActions.js.jsx
@@ -1,118 +1,148 @@
 const React = require("react");
 
-const Button         = require("../core/Button");
-const Form           = require("../core/Form");
-const HiddenFieldTag = require("../core/HiddenFieldTag");
-
 const ConfirmOrCancelBtns = require("../ConfirmOrCancelBtns");
 
 const ApprovedDeniedPendingBtnGroup = require("./ApprovedDeniedPendingBtnGroup");
+const ICalledButton                 = require("./ICalledButton");
 const PromptToCallTheBank           = require("./PromptToCallTheBank");
 
 const CallActions = React.createClass({
   propTypes: {
-    cardAccount: React.PropTypes.object.isRequired,
-    updatePath:  React.PropTypes.string.isRequired,
+    cardAccount:  React.PropTypes.object.isRequired,
+    submitAction: React.PropTypes.func.isRequired,
   },
 
+
   getInitialState() {
-    return { currentAction: "initial", }
+    return { currentState: "initial" };
     // States:
     // "initial", "called", "confirmApproved", "confirmPending", "confirmDenied"
   },
 
 
-  setCurrentAction(e, action) {
-    e.preventDefault();
-    this.setState({currentAction: action});
+  getAction() {
+    switch (this.state.currentState) {
+      case "confirmApproved":
+        return "call_and_open";
+      case "confirmDenied":
+        return "call_and_deny";
+      case "confirmPending":
+        return "call";
+      default:
+        throw "this should never happen";
+    }
+  },
+
+
+  getHelpText() {
+    switch (this.state.currentState) {
+      case "initial":
+        return `Please let us know when you've called ${this.bankName()}:`;
+      case "called":
+        return "What was the outcome?";
+      case "confirmApproved":
+        return "You were approved after calling for reconsideration:";
+      case "confirmDenied":
+        return "You called for reconsideration, but your application is still denied:";
+      case "confirmPending":
+        return "Your application is being reconsidered, and you're waiting to hear back:";
+    }
+  },
+
+
+  setStateToApproved() {
+    this.setState({currentState: "confirmApproved"});
+  },
+
+
+  setStateToCalled() {
+    this.setState({currentState: "called"});
+  },
+
+
+  setStateToCancel() {
+    this.setState({currentState: "called"});
+  },
+
+
+  setStateToDenied() {
+    this.setState({currentState: "confirmDenied"});
+  },
+
+
+  setStateToPending() {
+    this.setState({currentState: "confirmPending"});
+  },
+
+
+  submitAction() {
+    this.props.submitAction(this.getAction());
+  },
+
+
+  bankName() {
+    return this.props.cardAccount.card.bank.name;
+  },
+
+
+  phoneNumber() {
+    const bank = this.props.cardAccount.card.bank;
+    if (this.props.cardAccount.card.bp === "personal") {
+      return bank.personalPhone;
+    } else {
+      return bank.businessPhone;
+    }
   },
 
 
   render() {
     let buttons, helpText;
 
-    const bankName = this.props.cardAccount.card.bank.name
-
-    var action = "";
-
-    switch (this.state.currentAction) {
-      case "initial":
-        helpText = `Please let us know when you've called ${bankName}:`
-        break;
-      case "called":
-        helpText = "What was the outcome?"
-        break;
-      case "confirmApproved":
-        action = "call_and_open";
-        helpText = "You were approved after calling for reconsideration:"
-        break;
-      case "confirmDenied":
-        helpText = "You called for reconsideration, but your application is still denied:"
-        action = "call_and_deny";
-        break;
-      case "confirmPending":
-        helpText = "Your application is being reconsidered, and you're waiting to hear back:"
-        action = "call";
-        break;
-    }
-
-    switch (this.state.currentAction) {
-      case "initial":
-        buttons = (
-          <Button
-            primary
-            small
-            onClick={e => this.setCurrentAction(e, "called")}
-          >
-            I called {bankName}
-          </Button>
-        );
-        break;
-      case "called":
-        buttons = (
-          <ApprovedDeniedPendingBtnGroup
-            approvedText="I was approved after reconsideration"
-            deniedText="My application is still denied"
-            onClickApproved={e => this.setCurrentAction(e, "confirmApproved")}
-            onClickDenied={e => this.setCurrentAction(e, "confirmDenied")}
-            onClickPending={e => this.setCurrentAction(e, "confirmPending")}
-            pendingText="I'm being reconsidered, but waiting to hear back about whether it was successful"
-          />
-        )
-        break;
-      case "confirmApproved":
-      case "confirmDenied":
-      case "confirmPending":
-        buttons = (
-          <ConfirmOrCancelBtns
-            onClickCancel={e => this.setCurrentAction(e, "called")}
-            small
-          />
-        )
-        break;
-    }
-
-    const bank = this.props.cardAccount.card.bank
-    var   phoneNumber;
-    if (this.props.cardAccount.card.bp === "personal") {
-      phoneNumber = bank.personalPhone;
-    } else {
-      phoneNumber = bank.businessPhone;
-    }
-
     return (
-      <Form action={this.props.updatePath} method="patch">
-        <HiddenFieldTag name="card_account[action]" value={action} />
-
+      <div>
         <PromptToCallTheBank
           card={this.props.cardAccount.card}
           reconsideration
         />
 
-        <p>{helpText}</p>
+        <p>{this.getHelpText()}</p>
 
-        {buttons}
-      </Form>
+        {(() => {
+          switch (this.state.currentState) {
+            case "initial":
+              return (
+                <ICalledButton
+                  bankName={this.bankName()}
+                  onClick={this.setStateToCalled}
+                />
+              );
+            case "called": {
+              const pendingText = "I'm being reconsidered, but waiting to hear" +
+                                  " back about whether it was successful";
+              return (
+                <ApprovedDeniedPendingBtnGroup
+                  approvedText="I was approved after reconsideration"
+                  deniedText="My application is still denied"
+                  onClickApproved={this.setStateToApproved}
+                  onClickDenied={this.setStateToDenied}
+                  onClickPending={this.setStateToPending}
+                  pendingText={pendingText}
+                />
+              );
+            }
+            case "confirmApproved":
+            case "confirmDenied":
+            case "confirmPending":
+              return (
+                <ConfirmOrCancelBtns
+                  onClickCancel={this.setStateToCalled}
+                  onClickConfirm={this.submitAction}
+                  small
+                />
+              );
+          }
+        })()}
+      </div>
     );
   },
 });