+++ app/assets/javascripts/components/CardApplicationSurvey/CardApplicationSurvey.js.jsx
@@ -1,7 +1,13 @@
 const React = require("react");
+const _     = require("underscore");
+const humps = require("humps");
+const $     = require("jquery");
+
+const Alert = require("../core/Alert");
 
 const ApplyActions     = require("./ApplyActions");
 const CallActions      = require("./CallActions");
+const ExpiringText     = require("./ExpiringText");
 const NudgeActions     = require("./NudgeActions");
 const PostCallActions  = require("./PostCallActions");
 const PostNudgeActions = require("./PostNudgeActions");
@@ -9,17 +15,26 @@ const PostNudgeActions = require("./PostNudgeActions");
 const CardApplicationSurvey = React.createClass({
 
   propTypes: {
+    applyPath:   React.PropTypes.string.isRequired,
     cardAccount: React.PropTypes.object.isRequired,
-    updatePath:  React.PropTypes.string.isRequired,
+    declinePath: React.PropTypes.string.isRequired,
   },
+
+  getInitialState() {
+    return {
+      cardAccount: this.props.cardAccount,
+      ajaxStatus:  null,
+    };
+  },
+
 
   // Return the component that contains the actions for this card account.
   // Note that for some card accounts, no further action is required.
   getActionsComponent() {
-    const cardAccount = this.props.cardAccount;
+    const cardAccount = this.state.cardAccount;
 
     if (cardAccount.openedAt || cardAccount.redeniedAt) {
-      return undefined;
+      return "noscript";
     }
 
     if (!cardAccount.appliedAt) {
@@ -33,13 +48,69 @@ const CardApplicationSurvey = React.createClass({
     }
   },
 
+
+  submitAction(action, extraData) {
+    const data = {
+      _method: "patch",
+      "card_account[action]" : action,
+    };
+
+    if (action === "open" && extraData && extraData.openedAt) {
+      data["card_account[opened_at]"] = extraData.openedAt;
+    }
+
+    this.setState({ ajaxStatus: "loading"});
+
+    $.post(
+      `/api/v1/card_recommendations/${this.props.cardAccount.id}`,
+      data,
+      (response, textStatus, jqXHR) => {
+        if (response.error) {
+          this.setState({
+            ajaxStatus: "error",
+            ajaxError:  response.message,
+          });
+        } else {
+          const oldCardAccount = this.state.cardAccount;
+          // 'response' is the updated attributes of the card account
+          const newAttrs       = humps.camelizeKeys(response);
+          this.setState({
+            cardAccount: _.assign(oldCardAccount, newAttrs),
+            ajaxStatus:  "done",
+          });
+        }
+      }
+    );
+  },
+
+
   render() {
-    return React.createElement(
+    const actions = React.createElement(
       this.getActionsComponent(),
       {
-        cardAccount: this.props.cardAccount,
-        updatePath:  this.props.updatePath
+        applyPath:    this.props.applyPath,
+        cardAccount:  this.props.cardAccount,
+        declinePath:  this.props.declinePath,
+        submitAction: this.submitAction,
       }
+    );
+
+    return (
+      <div className="CardApplicationSurvey">
+        {actions}
+
+        {(() => {
+          switch (this.state.ajaxStatus) {
+            case "loading":
+              return <div className="LoadingSpinner" />;
+            case "done":
+              return <ExpiringText text="Saved!" />;
+            case "error":
+              return <Alert danger >{this.state.ajaxError}</Alert>;
+          }
+        })()}
+
+      </div>
     );
   },
 });