+++ spec/features/card_accounts/index_page/nudgeable_recs_spec.rb
@@ -1,7 +1,6 @@
 require "rails_helper"
 
 describe "user cards page - nudgeable cards", :js do
-
   include_context "logged in"
 
   let(:me) { account.owner }
@@ -23,9 +22,11 @@ describe "user cards page - nudgeable cards", :js do
     )
     visit card_accounts_path
   end
+
   let(:rec)  { @rec }
   let(:bank) { @bank }
-  subject(:rec_on_page) { NudgeableCardAccountOnPage.new(rec, self) }
+
+  let(:rec_on_page) { NudgeableCardAccountOnPage.new(rec, self) }
 
   shared_examples "clicking 'cancel'" do
     describe "and clicking 'cancel'" do
@@ -41,19 +42,17 @@ describe "user cards page - nudgeable cards", :js do
     end
   end
 
-  it "says when I applied", :frontend do
+  example "nudgeable rec on page", :frontend do
     expect(rec_on_page).to have_content "Applied: #{applied_at.strftime("%D")}"
-  end
-
-  it "doesn't have apply/decline or 'I applied' buttons", :frontend do
+    # has buttons:
+    expect(rec_on_page).to have_i_called_btn
+    expect(rec_on_page).to have_i_heard_back_btn
     expect(rec_on_page).to have_no_apply_btn
     expect(rec_on_page).to have_no_decline_btn
     expect(rec_on_page).to have_no_i_applied_btn
-  end
-
-  it "encourages me to call the bank", :frontend do
-    is_expected.to have_content "We strongly recommend that you call #{bank.name}"
-    is_expected.to have_content(
+    # it encourages me to call the bank:
+    expect(rec_on_page).to have_content "We strongly recommend that you call #{bank.name}"
+    expect(rec_on_page).to have_content(
       "You’re more than twice as likely to get approved if you call #{bank.name} "\
       "than if you wait for them to send your decision in the mail"
     )
@@ -62,23 +61,17 @@ describe "user cards page - nudgeable cards", :js do
   context "for a personal card" do
     let(:bp) { :personal }
     it "gives me the bank's personal number" do
-      is_expected.to have_content "call #{bank.name} at 888-245-0625"
-      is_expected.to have_no_content "800 453-9719"
+      expect(rec_on_page).to have_content "call #{bank.name} at 888-245-0625"
+      expect(rec_on_page).to have_no_content "800 453-9719"
     end
   end
 
   context "for a business card" do
     let(:bp) { :business }
     it "gives me the bank's business number" do
-      is_expected.to have_content "call #{bank.name} at 800 453-9719"
-      is_expected.to have_no_content "888-245-0625"
+      expect(rec_on_page).to have_content "call #{bank.name} at 800 453-9719"
+      expect(rec_on_page).to have_no_content "888-245-0625"
     end
-  end
-
-
-  it "has buttons to say I called or I heard back", :frontend do
-    expect(rec_on_page).to have_i_called_btn
-    expect(rec_on_page).to have_i_heard_back_btn
   end
 
   describe "clicking 'I called'" do
@@ -121,20 +114,16 @@ describe "user cards page - nudgeable cards", :js do
       describe "and clicking 'confirm'" do
         before do
           rec_on_page.click_confirm_btn
+          # FIXME can't figure out a more elegant solution than this:
+          sleep 1.5
           rec.reload
         end
 
-        it "marks the rec as 'open'", :backend do
+        it "updates the rec's attributes", :backend do
           expect(rec.status).to eq "open"
-        end
-
-        it "sets 'opened_at' and 'nudged_at' to the current date", :backend do
           expect(rec.opened_at).to eq Date.today
           expect(rec.nudged_at).to eq Date.today
-        end
-
-        it "doesn't change applied_at", :backend do
-          expect(rec.applied_at).to eq applied_at
+          expect(rec.applied_at).to eq applied_at # unchanged
         end
       end
     end
@@ -147,18 +136,14 @@ describe "user cards page - nudgeable cards", :js do
       describe "and clicking 'confirm'" do
         before do
           rec_on_page.click_confirm_btn
+          # FIXME can't figure out a more elegant solution than this:
+          sleep 1.5
           rec.reload
         end
 
-        it "marks the rec as 'denied'", :backend do
+        it "updates the rec's attributes", :backend do
           expect(rec.status).to eq "denied"
-        end
-
-        it "doesn't change 'applied_at'", :backend do
-          expect(rec.applied_at).to eq applied_at
-        end
-
-        it "sets 'denied_at' and 'nudged_at' to the current time", :backend do
+          expect(rec.applied_at).to eq applied_at # unchanged
           expect(rec.denied_at).to eq Date.today
           expect(rec.nudged_at).to eq Date.today
         end
@@ -173,15 +158,14 @@ describe "user cards page - nudgeable cards", :js do
       describe "and clicking 'confirm'" do
         before do
           rec_on_page.click_confirm_btn
+          # FIXME can't figure out a more elegant solution than this:
+          sleep 1.5
           rec.reload
         end
 
-        it "doesn't change the rec's status or 'applied_at' timestamp", :backend do
+        it "updates the rec's attributes", :backend do
           expect(rec.status).to eq "applied"
-          expect(rec.applied_at).to eq applied_at
-        end
-
-        it "sets 'nudged_at' to the current time", :backend do
+          expect(rec.applied_at).to eq applied_at # unchanged
           expect(rec.nudged_at).to eq Date.today
         end
       end
@@ -210,11 +194,8 @@ describe "user cards page - nudgeable cards", :js do
     end
 
     shared_examples "doesn't change applied or nudged" do
-      it "doesn't set 'nudged_at'", :backend do
+      it "doesn't change applied or set nudged", :backend do
         expect(rec.nudged_at).to be_nil
-      end
-
-      it "doesn't change applied_at", :backend do
         expect(rec.applied_at).to eq applied_at
       end
     end
@@ -236,14 +217,13 @@ describe "user cards page - nudgeable cards", :js do
       describe "and clicking 'confirm'" do
         before do
           rec_on_page.click_confirm_btn
+          # FIXME can't figure out a more elegant solution than this:
+          sleep 1.5
           rec.reload
         end
 
-        it "marks the rec as 'open'", :backend do
+        it "updates the rec's attributes", :backend do
           expect(rec.status).to eq "open"
-        end
-
-        it "sets 'opened_at' to the current date", :backend do
           expect(rec.opened_at).to eq Date.today
         end
 
@@ -259,14 +239,13 @@ describe "user cards page - nudgeable cards", :js do
       describe "and clicking 'confirm'" do
         before do
           rec_on_page.click_confirm_btn
+          # FIXME can't figure out a more elegant solution than this:
+          sleep 1.5
           rec.reload
         end
 
-        it "marks the rec as 'denied'", :backend do
+        it "updates the rec's attributes", :backend do
           expect(rec.status).to eq "denied"
-        end
-
-        it "sets 'denied_at' to the current date", :backend do
           expect(rec.denied_at).to eq Date.today
         end